chore: move private headers
