+++ src/unix/sdl_render.c
@@ -18,8 +18,9 @@
 #include <86box/version.h>
 #include <86box/ini.h>
 #include <86box/config.h>
-#include <86box/unix_osd.h>
-#include <86box/unix_sdl.h>
+
+#include "sdl_render.h"
+#include "sdl_osd.h"
 #include "sdl_shader.h"
 
 #define RENDERER_FULL_SCREEN 1
